完成排行榜

+++ src/components/AdminViewComponents/AdvertiseContentEl.vue
@@ -97,16 +97,37 @@
 
       },
       del(){
-        console.log("删除")
-        var url="http://localhost:8080/onlineBookStore/advertise/"+this.element.advId
-        this.$http.delete(
-          url
-        ).then(response=>{
-          console.log(response.data)
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          var url="http://localhost:8080/onlineBookStore/advertise/"+this.element.advId
+          this.$http.delete(
+            url
+          ).then(response=>{
+            console.log(response.data)
 
-          this.$router.go(0)
+            this.$router.go(0)
 
-        })
+          })
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          });
+        });
+
+        // console.log("删除")
+        // var url="http://localhost:8080/onlineBookStore/advertise/"+this.element.advId
+        // this.$http.delete(
+        //   url
+        // ).then(response=>{
+        //   console.log(response.data)
+        //
+        //   this.$router.go(0)
+        //
+        // })
       },
       toFade() {
         this.isReal = false
